Refactor initial data loading + improve it
I added a single method that loads data, and it is called for each of the file. Also the bug with the "NA" trying to be saved as a number is naively solved, but it should be solved in a better way (check for attribute type, replace with undefined only if it's a number).